[NEEDS-CLEANUP] nvmeof: initialize global RBD variables
The NVMe-oF controller uses the RBD controller, so it should initialize
all global RBD variables too (yuck).

remove ONLINE VolumeExpansion for now
This should work the same way as RBD supports it. The online part needs
support in the node-plugin as well.

Initial work focusses on CreateVolume and DeleteVolume. Additional
features will be implemented afterwads.

Signed-off-by: gadi-didi <gadi.didi@ibm.com>